refactor: move spec files into __tests__ directories
