+++ packages/sheets-crosshair-highlight/src/controllers/__tests__/crosshair.controller.spec.tsx
@@ -21,11 +21,11 @@ import {
     EnableCrosshairHighlightOperation,
     SetCrosshairHighlightColorOperation,
     ToggleCrosshairHighlightOperation,
-} from '../commands/operations/operation';
-import { CROSSHAIR_HIGHLIGHT_OVERLAY_COMPONENT } from '../menu/crosshair.menu';
-import { menuSchema } from '../menu/schema';
-import { CrosshairOverlay } from '../views/components/CrosshairHighlight';
-import { SheetsCrosshairHighlightController } from './crosshair.controller';
+} from '../../commands/operations/operation';
+import { CROSSHAIR_HIGHLIGHT_OVERLAY_COMPONENT } from '../../menu/crosshair.menu';
+import { menuSchema } from '../../menu/schema';
+import { CrosshairOverlay } from '../../views/components/CrosshairHighlight';
+import { SheetsCrosshairHighlightController } from '../crosshair.controller';
 
 describe('SheetsCrosshairHighlightController', () => {
     it('should register commands, merge menu, register components and register icons', () => {
