+++ packages/sheets-crosshair-highlight/src/views/widgets/__tests__/crosshair-highlight-shape.spec.ts
@@ -15,7 +15,7 @@
  */
 
 import { describe, expect, it } from 'vitest';
-import { SheetCrossHairHighlightShape } from './crosshair-highlight-shape';
+import { SheetCrossHairHighlightShape } from '../crosshair-highlight-shape';
 
 class TestCanvasContext {
     fillStyle = '';
