+++ packages/watermark/src/facade/__tests__/f-univer.spec.ts
@@ -19,7 +19,7 @@ import { FEnum, FUniver } from '@univerjs/core/facade';
 import { IWatermarkTypeEnum } from '@univerjs/engine-render';
 import { WatermarkService, WatermarkTextBaseConfig } from '@univerjs/watermark';
 import { describe, expect, it } from 'vitest';
-import { FUniverWatermarkMixin, FWatermarkEnumMixin } from './f-univer';
+import { FUniverWatermarkMixin, FWatermarkEnumMixin } from '../f-univer';
 
 class TestWatermarkService {
     readonly updatedConfigs: unknown[] = [];
